+++ src/Symfony/Component/Security/Http/RememberMe/RememberMeHandlerInterface.php
@@ -21,6 +21,10 @@ use Symfony\Component\Security\Core\User\UserInterface;
  * {@see AbstractRememberMeHandler} instead.
  *
  * @author Wouter de Jong <wouter@wouterj.nl>
+ * @author Patrick Elshof <tyrelcher@protonmail.com>
+ *
+ * @method string            getUserIdentifierForCookie(RememberMeDetails $rememberMeDetails) Retrieves the User Identifier for the RememberMe cookie.
+ * @method RememberMeDetails getRememberMeDetails(string $rawCookie)                          Retrieves the RememberMeDetails using the raw cookie.
  */
 interface RememberMeHandlerInterface
 {